✨ Add semicore support for PseudoDojo v0.5 PBE
Generate the JSON files indicating the semicore orbitals for PseudoDojo v0.5
PBE, including the `standard` and `stringent` protocols.

To show the efficacy of the script used to generate the JSON files, the JSONs
containing the semicores for PseudoDojo v0.4 LDA/PBE standard have also been
regenerated.

Additionally, the "stringent" versions of the PseudoDojo v0.4 PBE/LDA versions
are also generated.

To test that all pseudo potentials are truly supported, a sanity test is added
which checks that the `get_pseudo_orbitals()` function can find all pseudos for
the supported families.